Moved CellWorldObjectDrawer, ColoringStrategy, GameFX into new package it.cwmp.client.view.game.drawing
