+++ client/src/main/scala/it/cwmp/client/view/game/drawing/GameFX.scala
@@ -1,7 +1,8 @@
-package it.cwmp.client.view.game
+package it.cwmp.client.view.game.drawing
 
 import akka.actor.ActorRef
 import it.cwmp.client.model.game.impl.{Cell, CellWorld, Point}
+import it.cwmp.client.view.game.GameViewActor
 import it.cwmp.client.view.game.model.{CellView, TentacleView}
 import it.cwmp.client.view.{FXAlertsController, FXViewController, LayoutRes}
 import javafx.fxml.FXML
